+++ CHANGELOG.md
@@ -4,7 +4,18 @@ This changelog follows the specifications detailed in: [Keep a Changelog](https:
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html), although we have not yet reached a `1.0.0` release.
 
 
-## Version 0.8.0 - Unreleased
+## Version 0.8.1 - Unreleased
+
+### Added
+
+New convenience arguments to `.cli`
+
+
+## Version 0.8.0 - Released 2024-08-14
+
+### Add
+* Add experimental new flag `__allow_newattr__` which relaxes the constraint
+  that you can't add keys on the fly.
 
 ### Removed